requirements.txt
pip==24.2
Django==5.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-nested-routers==0.93.5
django-extensions==3.2.3
load_dotenv==0.1.0
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.35
langchain_openai==0.1.22
python-dotenv==1.0.1
streamlit==1.37.1
pymysql==1.1.1
pytz~=2024.1
PyJwt==2.9.0
gunicorn==23.0.0
faiss-cpu